Changed class dependency for value fields
Fixes #302
